New File, New Idea, New Dimension
Changed screen size dimension for MenuUI because of mac screen size
bug, allowed an additional 50px for height.
Created a separate file for returning printer results in a different
JFrame.
Intention is for user to click filter results, and have a entirely new
window pop up with refined results based on weighted input (yet to be
implemented).

As a matter of fact, the window won’t open at all, and it’s pissing me
off. I would like to sit down with others to work on this sometime this
week to get it responsive so we have the action listener ready to
initiate the pop up window based on user parameter search entries.