Printing Feature
Added the feature of allowing the Printer to print the results.
Actually retrieving the results for printing will be later implemented,
the image thats printed is a 2D graphic.
Retrieving the values will simply retrieve textfields from the desired
results, and have hard coded formatting for the pdf to print as wanted!
For now the print should work, and allow you to save as pdf, it’ll save
a mini text that says “TREVOR I KNOW WHAT YOU DID LAST SUMMER”.