Reformatted Javadoc
I noticed that the printer method JavaDoc was separated from the method
and wasn't sure if it was due to the length.  I ended up moving the
minimal information for the JavaDoc back to the printer method and left
the notes in the class JavaDoc.

In addition, I moved class descriptions above the author tag.