Printing compatible on both devices now
Made printing compatible on both devices, also revised error situation
for try catch block instructing application to be restarted in
erroneous situation of attempting to print.
Currently works on Mac, should work equally the same on Windows.
POSSIBLE ERROR: to check for is printing a list that is twice as long
as the one we currently have!